Bug 881: Bring Wi-Fi IE serdes intelligence into WifiInformationElement object
Introduce WifiInformationElement methods SerializeIE(),
DeserializeIE(), and DeserializeOptionalIE() (the latter for use when
the IE of interest might not be present), which know how to deal with
the IE as a whole.

Make use of these in mesh subsystem where WifiInformationElement-derived
objects are used.

Also add empty implementation of Print() method of
WifiInformationElement so that IEs that don't care about being
displayed don't need to implement this.