Add "unfixable" ByteTagList test case
This test case creates empty packet, removes last 50 bytes and
adds them back.

As packet is empty initially, no bytes are marked dirty.  Therefore,
Buffer::AddAtEnd returns false, indicating that added bytes were not used
before.  However, it does not know if they were tagged with byte tags.

This bug is not so easy to fix as Buffer does not have enough information
to return true in Buffer::AddAtEnd when it should.