wifi: (partial fix #1797) Align legacy Minstrel with Linux and minstrel_ht
The changes in Minstrel are inteded to follow more accurately the Linux
implementation.

For each rate, a new parameter samplesSkipped is added. This parameter is
intended to solve an issue regarding the sampling of low rates when a high
rate is working well, which leads to outdated statistics.
See [1] for further explanation.

This change makes throughput a bit lower in simple, stable scenarios. However,
it may help in dynamic scenarios to react faster and more accurately to
changes.

Related to the previous, the logic is changed for deciding when to sample
random rates. When a sample rate is deferred to the second MRR chain stage,
a new parameter (numSamplesDeferred) is increased. This parameter is used
(jointly with sampleCount) to compare current sample count with the
lookaround rate.

Also related with sampling, another parameter is added, sampleLimit.
This parameter limits the number of times a very low or very high
probability rate is sampled, avoiding to try a poorly working sample rate
too often.

When updating the EWMA probability of a rate for the first time, do not
apply EWMA but assign the entire probability.  As EWMA probability is
initialized with zero, this generates a more accurate EWMA.

[1] http://thread.gmane.org/gmane.linux.kernel.wireless.general/104920/focus=104923